Changed paramater types of Multiply Method

+++ GitPractice/Calculator.cs
@@ -30,7 +30,7 @@ namespace GitPractice
             return num1 - num2;
         }
 
-        public int Multiply(int num1, int num2)
+        public double Multiply(double num1, double num2)
         {
             return num1 * num2;
         }
